refactor: move LobbyLayout to src/layout/ alongside components, stores, composables
